+++ test/test152.js
@@ -5,7 +5,10 @@ if(typeof exports === 'object') {
 	__dirname = '.';
 };
 
-if(typeof exports != 'object') {
+if(typeof exports == 'object') {
+	var DOMStorage = require("dom-storage")
+	global.localStorage = new DOMStorage("./test149.json", { strict: false, ws: '' });
+};
 
 describe('Test 152 - INSERT/DELETE/UPDATE for localStorage with AUTOCOMMIT', function() {
 
@@ -57,5 +60,4 @@ describe('Test 152 - INSERT/DELETE/UPDATE for localStorage with AUTOCOMMIT', fun
 	});
 });
 
-}