+++ test/test182.js
@@ -7,7 +7,7 @@ if(typeof exports === 'object') {
 
 //if(typeof exports != 'object') {
 
-describe('Test 181 - ARRAY aggregator', function() {
+describe('Test 182 - ARRAY aggregator', function() {
 
 	it("1. ARRAY()", function(done) {
 var data = [
@@ -53,7 +53,8 @@ var data = [
       ARRAY({category:category,[count]:[count]}) AS purchases, SUM([count]) AS totalCount \
       FROM ? GROUP BY userId, userName',[data]);
  
-     console.log(res);
+    assert(res.length == 2);
+//     console.log(res);
 //      assert.deepEqual(res,[1,2,3,4,5,6,7,8,9,10]);
       done();
   });