+++ test/test196.js
@@ -18,14 +18,14 @@ describe('Test 196 - COUNT(a) vs COUNT(*)', function() {
             { a: 2, b: 2, c: 2 },
             { a: 3, b: 1, c: 1 },
             { a: undefined, b: 1, c: 0 } ]);
-        console.log(res);
+//        console.log(res);
     	done();
     });
     
     it("2. COUNT(DISTINCT a)", function(done) {
         var data = [{a:1},{a:1},{a:2},{a:3},{a:1},{a:2}];
         var res = alasql('SELECT COUNT(DISTINCT a) FROM ?',[data]);
-        console.log(res);
+//        console.log(res);
         done();
     });
 });